Comment out non-functional unit test
This unit test will corrupt the heap each time it runs since it's trying to deallocate memory that was allocated in HumanFactors.dll. I've added code to show users how to perform the deletion, but it shouldn't be done executed outside of the humanfactors DLL.

The C_Interface provides functions  for external callers  to deallocate this memory,  so testing for InsertPathsIntoArray should instead be tested there.