Added missing examples for RayRequest, tested in tests/src/embree_raytracer.cpp
The structures/functions in RayRequest.h that were marked as deprecated were
skipped for this pass -- if a structure/method/function was not marked as
deprecated, it was given an example.

Since there was no dedicated RayRequest test source file,
the tests for RayRequest were placed in test/src/embree_raytracer.cpp.

There were two items tested - the constructor (with args) and DidHit.
Those items, plus the no-arg constructor were given example code in
RayRequest.h.