Work done so far with cross slope.
Within the for-loop (lines 59-78 in cost_algorithms.cpp)
is the traversal of the CSR. Parent IDs and child IDs are
extracted from the CSR, along with the edge data relevant to the
parent ID/child ID pairing.

parent_id, child_id, and *edge_data can be used to build an IntEdge,
(after the cross slope is calculated, of course)
which will be added to a container that is returned from this function.