First version of cross slope algorithm
Still needs different types of graphs to verify that this works as intended,
but CSR traversal works as intended.

From this point on, it is a matter of g.NodesFromID working to get
the nodes from the graph, in order to do the dot product calculations
(i.e. if dot product of two vectors is 0, we have a perpendicular edge)