Further revision to CalculateCrossSlope
The use of the CSR no longer makes sense for this function --
since cross slope is highly dependent on position data,
thus we require Node and Edge instances.

With the added functionality in Graph,
it is easy to retrieve all children for a given parent node,
or even its parent id.

CalculateCrossSlope now just needs to be tested for unique cases,
but the algorithm is coming together.