Updates to CalculateCrossSlope, and added CalculateCrossSlopeCSR
CalculateCrossSlopeCSR encorporates portions of my former implementation
of CalculateCrossSlope.

The virtues of going forward with the CSR variant are as follows:
In CalculateCrossSlope (original),
we load up all the parent nodes from the graph at once,
even though we only need one parent node at at time.

In CalculateCrossSlopeCSR,
we use a for-loop to iterate from [0, csr.rows), which are all
the parent IDs. For each parent ID, I call
g.NodeFromID(parent_id) -- which is the same as
g.ordered_nodes[parent_id]. I also mark the beginning of a row,
and the end of a row -- so that I know which child IDs belong
to the current parent ID.

Once I have the desired parent id and child id (and of course
getting the Node object from the child id) -- I get all of the edges
that extend from parent_id, and call
g.GetPerpendicularEdges(parent_node, edges, child_node)

Then, the rest remains the same as CalculateCrossSlope so far.

It is hard to say which version will win out performance wise,
as more testing still needs to take place -- I do prefer the notational
conveniences of CalculateCrossSlope (the original), as it appears to be
far more maintainable and easier to follow -- but the CSR variant
may have a performance advantage, simply because the parent nodes
need not be loaded at once.