CInterface implementation for CalculateCrossSlope (CSR variant)
For now, I will be moving forward with the CSR version of
CalculateCrossSlope, as it may prove more efficient later on
(since we need not load all nodes in a graph into memory at once)

CInterface implementation for CalculateAndStoreCrossSlope works as follows:
- get a std::vector<Edge> by calling CalculateCrossSlopeCSR on g
- get a ptr to underlying buffer inside the std::vector<Edge>
- retrieve the CSR representation of g
- then, for all parent id [0, csr.nnz),
- get the addresses for the beginning and end of a row within the CSR,
since these addresses mark the boundaries of what child ID addresses belong
to the current parent_id. (row_curr and row_end)
- then, while row_curr < row_end, calculate desired weight. More below:

- The 'desired_weight' is the difference between the cross slope value and the current edge value.
- It seems that Graph::addEdge(int, int, float) adds the float parameter with the current edge value,
which is not what we want -- we want the cross slope values to be THE edge values for all edges when
using CalculateAndStoreCrossSlope. This difference will account for that problem,
but more investigating will be required.

- Compress the graph after adding all edges.

- The end result of CalculateAndStoreCrossSlope is the same Graph g,
but with the cross slope data calculated from CalculateCrossSlopeCSR,
replacing the former edge data.