First implementation of C++ code for iss 31 - Node attributes
A new field, named attr_map, was added to Graph. It is:
robin_hood::unordered_map<std::string, Eigen::SparseMatrix<float, 1>>

The key is an attribute type, like "default", "cross slope", "energy expenditure", etc.
The value is the unique CSR representation of the Graph, given that attribute type.

The member functions in iss 31 are not responsible for creating these CSRs,
they are only responsible for adding/retrieving/removing data from existing
CSRs, that were added using the member functions described in issue 24 -
Multiple Costs for the Graph. Note that issue 24 has not been completed yet.

The most important addition was the inclusion of the attr_map field,
as this is where mappings of node attribute types to CSRs will be stored.