Node attributes (C++ portion) redone
Disregard previous commits regarding this topic,
this is being redone. The CInterface commit that corresponds
to this feature will follow.

A new field has been added in Graph,
std::unordered_map<std::string, NodeAttributeValueMap> node_attr_map;

NodeAttributeValue is an identifier from this using statement:
using NodeAttributeValueMap = std::unordered_map<int, std::string>

In node_attr_map[attribute_name],
the key is the attribute_name, ie "cross slope", "energy expenditure", etc.
the value is a NodeAttributeValueMap for all data related to attribute name.

In a NodeAttributeValueMap,
the key is a node id, and the value is a quantity, or figure, for a node attribute,
in the form of a std::string. So far,
valid types are either string types, or floating point types.
This code can differentiate between the two, via a new function
called is_floating_type, in the CostAlgorithms namespace.