Removed non-csr version of CalculateCrossSlope. Removed/relabled commented-out test cases
Since I am moving forward with the CSR version of CalculateCrossSlope,
I have removed the non-CSR version of said function --
and have renamed CalculateCrossSlopeCSR to be CalculateCrossSlope
(since it is the only function of its kind now)

Also relabeled the test cases that have been commented out --
GraphGenerator needs to be fixed in order for the tests using an OBJ
file can be run - GraphGenerator throws an exception for energy_blob.obj.
Until this is fixed -- I have provided tests with 'handwritten' graphs
for CalculateCrossSlope and CalculateEnergyExpenditure.