Add some new functions for resizing compressed CSRs
At the moment, we likely don't want to do this much, since it will be a reallocation for each new node in the graph, however it may become a good option if we can preallocate the array size. 

This may be an attractive option for alternate costs since the size of the graph only needs to change with the size of ordered_nodes, so alternate costs may be able to skip this penalty.