Decouple node ID from CSR Rows
Previously, nodes were stored in order of ID with each ID corresponding to a different row in the CSR. This caused issues when trying to add integer edges to the graph and overall made it more complicated to make bigger changes to the structure of the graph. 

Moving forward, each node will have an index, and an ID. The ID is the unique identifier assigned to the node that can (soon) be specified by the user, then the index is the node's index in the ordered_nodes array and the row number of the node in the CSR. 

I've included a new testcase to test this functionality, and more are to come.