Added bodies for CreatePath/Paths (for cost_name) for C++/CInterface
Added bodies for the following:
- In C++
- std::vector<EdgeSet> Graph::GetEdges(const string cost_name) const;
- BoostGraph::BoostGraph(const string cost_name);
- std::unique_ptr<BoostGraph, BoostGraphDeleter> CreateBoostGraph(const Graph& g, const string cost_name);

- In C_Interface
- C_INTERFACE CreatePathCostName(...same as CreatePath, const char *cost_name);
- C_INTERFACE CreatePathsCostName(...same as CreatePaths, const char *cost_name);

These methods/functions have bodies for testing in the respective test source files.

Graph::GetEdges is to be implemented from issue 24, Multiple Costs for the Graph.
Searching for cost_name will require HF::Exceptions::NoCost (from issue 24)
if an argument cost_name is not found in a Graph g.