C Interface functions, CreatePathCostType and CreatePathsCostType, with tests
CreatePathCostType and CreatePathsCostType share nearly the same
prototypes as that of CreatePath and CreatePaths respectively --
with a new parameter at the end of each: const char *cost_name.

These functions use nearly the same implementation as the
non-cost type variants, but since CreateBoostGraph(Graph&, std::string)
can throw a std::out_of_range exception (which originates from
Graph::GetEdges(std::string cost_name), and is called in
BoostGraph::BoostGraph(Graph&, string), which is called in CreateBoostGraph(Graph&, std::string),
the implementation for each accomodates the possibility of an exception being thrown,
and when (if) the exception is caught -- HF::Exceptions::HF_STATUS::GENERIC_ERROR
is the error code return.

GENERIC_ERROR should really be NO_COST, but that will be left
to issue #24, 'Multiple Costs for the Graph'.

The functions in this branch cannot be tested
 without the complete implementation of Graph::GetEdges(std::string cost_name),
and the hashmap field in Graph (cost_type_map, or however it will be named --
which is a robin_hood::unordered_map<std::string, Eigen::SparseMatrix<float, 1>>).

(We also need a method for adding typed edges to the graph, which is also to be implemented
from #24)