Newly set up project with explicit loading of DLL. Still need to load remaining functions.
This is a newly set up project with an x64-Release configuration,
using explicit loading of the DLLs required by the Human Factors C Interface API.

As of now, the remainder of the functions for the raycasting example have to be loaded,
but the LoadOBJ function has been loaded, and serves as a 'prototype' for the remaining
functions to load.

After the functions are loaded, the example can continue as it once was
in previous iteratons of this project.

Each example will have its own .cpp file.