Raycasting example and updated client source. Trying to resolve exception thrown from FreeLibrary (see description)
test_raycasting.cpp has the Raycasting example, where functions are loaded
from the DLL. Output matches that of the Python example (it is the same code as before)
but when calling FreeLibrary(dll_human_factors), an exception is thrown.

However, when using the debugger to step through the code
(especially at that problematic if (FreeLibrary(dll_human_factors)) statement),
FreeLibrary(dll_human_factors) runs just fine.

This is at the end of the program, but I would still like to resolve this problem.

FreeLibrary(dll_tbb) and FreeLibrary(dll_embree3) work fine.