Found a fix for FreeLibrary(dll_humanfactors). Sleep the current thread for 250ms before calling FreeLibrary
The FreeLibrary(dll_humanfactors) statement works when stepping through the code
with the debugger, but ends up throwing an exception when the program is run.

By putting the current thread to sleep for 250 ms (before calling FreeLibrary(dll_humanfactors)),
FreeLibrary is now able to work with dll_humanfactors.

I have provided the link as to where I found this suggested technique.
(there is likely some 'housekeeping' occuring by dll_humanfactors?
Will have to study this more to have a better answer as to the reason
why this is necessary)