Assert correctness of Pathfinding examples using string comparsions
The output that's printed to console by the tests in examples is compared to a copy of the expected output in the documentation at the time of writing. 

Other examples should continue to use assertions, as they are much more informative and robust than this method, however it is not practical for examples with this volume of output.