Example for loading DLL at top of file, and example for CreateRaytracer
At the top of this file, we describe the code required
to load the HumanFactors DLL (and the DLLs it depends on).

Each example will assume this code has run, prior to any of the
C Interface functions being used.

Then, the CreateRaytracer example has been added.
Before the example itself, we describe how to obtain access
to the C Interface functions (via the function pointers returned by GetProcAddress).

Then the CreateRaytracer example is shown.

All examples will follow this procedure.