Converted samples with \code to \snippet. Snippets drawn from unit test source file (see description)
Since all of our examples needed to be tested anyway (and demonstrated to work via the Test Explorer),
the snippets can be drawn from Analysis/src/Cpp/tests/src/<containing file>.

The files where the snippets are being drawn from are already part of the codebase
(which is what the \snippet tag is primarily designed to work with) --
so this prevents the need to have a separate file for the sole purpose of
storing snippets to be displayed for documentation.

Since our EXAMPLE_PATH variable in our C++ Doxyfile is set to
'Cpp', the path specified was

tests/src/embree_raytracer_cinterface.cpp

embree_raytracer_cinterface.cpp would be replaced with whatever
test source file that contains the snippet.

Full syntax in raytracer_C.h for invoking an existing snippet is,
say, for the CreateRaytracer function is this:

\snippet tests\src\embree_raytracer_cinterface.cpp snippet_CreateRaytracer

Then, in the test source file,
//! [snippet_CreateRaytracer]
code here
//! [snippet_CreateRaytracer]