Fully configured Visual Studio solution (.sln) with projects (.vcxproj) inside the solution representing each example for Sample Usage
This was the original configuration of the .sln
containing .vcxproj files for all of examples in the 'Sample Usage' series.

All .vcxproj files are configured to have one build mode: x64-Release,
and they are also configured with the C++ version: ISO C++17.

The sources have not been modified.

Note: Since 'Visualize the Graph' has not been merged to master,
it has been excluded from here. When 'Visualize the Graph' is merged
to master, it will be added to the .sln
with a configured .vcxproj file to go with it.