New section at top of raytracer_C.h: Raytracer setup/teardown. Reduction of repeated code in snippets
All of the examples begin and end with the same 'setup' and 'teardown'
code. At the top of raytracer_C.h (in the 'Detailed Description' section of Raytracer documentation),
I have added two sections: Raytracer setup and Raytracer teardown.

Because each example will begin with 'Raytracer setup',
and each example will end with 'Raytracer teardown',
it is better to reduce code repetition and have the reader review
the setup/teardown sections (every example will have a \see tag, for See also - Raytracer setup, Raytracer teardown)

As a result, the snippets only involve the setup/teardown that is
unique to that example. Raytracer setup and teardown are common to
all of the examples, so by 'factoring out' that code, we can reduce
the page clutter overall.

I still want to find a way to link 'Raytracer setup' and 'Raytracer teardown'
so they can be clicked -- and the user can immediately see what that
section says, if they want to read it.