Added new section: Notes on PerformRotation, revised hyperlinks
Notes on MeshInfo::PerformRotation mentions that LoadOBJ and RotateMesh
contain a call to HF::Geometry::MeshInfo::PerformRotation - which has
a body of code containing assert statements that ensure that no
NANs or infinity values were created from the rotation values passed
in by the caller.

If the assert statements evaluate false (meaning that there was an occurence
of a NAN and/or infinity value, the program will halt execution)

Being that this applies to both LoadOBJ and RotateMesh,
I did not want to have to provide this same 'warning' twice --
so I made a section for it in 'Detailed Description'.