Updated end of description for Notes on PerformRotation. Needed to be more clear.
"These assertion statements may evaluate false (which will halt execution)
if NANs (not-a-number) or infinity values were created
from the xrot, yrot, or zrot values passed to LoadOBJ or RotateMesh."

The previous version of this statement ended with
"passed to this function." What function?

(The Notes on PerformRotation remark was originally placed with the
LoadOBJ and RotateMesh documentation, which explains the phrasing of
'this function'. Now that this entry is in the 'Detailed description'
section, I needed to be more clear as to what functions I was talking about)