Examples for SphericalViewAnalysisNoAggregateFlat and SphereicalViewAnalysisAggregateFlat
These snippets were derived from the Calculating Spatial View example
from Sample Usage. Note that the \ref tags to raytracer_setup
and raytracer_teardown will not work until the documentation revisions for
raytracer_C are available. (The \sections that the \ref tags use are on the docs_cinterface_raytracer_gem branch,
but I do not want to mix that up with the work on this branch.)