Added guard statement for header file
General rule of thumb:
Always add include guards
(the traditional style, #ifndef DIRECTIVE, #define DIRECTIVE // code, #endif /* DIRECTIVE */)
to avoid duplicate declarations/definitions of namespaces/types/functions/etc.
(#pragma once) is an alternative way to do this, but it is not supported by all compilers.

It IS supported by the Visual C++ compiler (what we are using in this repo),
so it's fine on that front -- but I find it safer to use the traditional guard style.