Added section for finding closest nodes to specified points.
We do not have a 'get_closest_nodes' function in the C Interface,
so I took the euclidean_distance template I wrote from cost_algorithms.h
and put it in this translation unit. Then I manually computed the
closest nodes to the specified points using an iterator from node_vector.