Wrote a more generalized function for getting the closest nodes.
The C Interface function doesnt have a function that will return a container
of the nodes in a graph that have the shortest distance from a set of comparison nodes.
The function is specific to the example, but it can be reworked so that it can
possibly become part of the C Interface.