Added \ref tags for mesh_setup and mesh_teardown (will refer to sections in objloader_C.h)
The \section tags that the \ref tags refer to are in objloader_C.h,
in the docs_cinterface_objloader_gem branch. When all doc revision branches
are merged together, the \ref tags will resolve normally.

As of now, all \ref tags will appear verbatim
(i.e. the name of the section handle displays, instead of a link to the section itself)