Reworked snippets for FireMultipleOriginsOneDirection to hide GTEST ASSERT statements.
All of the other C Interface revision branches follow the same principle of "breaking up"
or "separating" the example by splitting it up into several snippets (that are part of the same whole)

This is how we can get around hiding the ASSERT statements throughout the test sources.

Breaking up the example also helps so things don't look too cluttered.