Added section tags in Detailed Description
The functions CreateGraph, AddEdgeFromNodes, AddEdgeFromNodeIDs, Compress, GetCSRPointers, and DestroyGraph are used in several examples. The use of these functions can be placed at the top of Detailed Description, where we walk through the process of creating a graph from objects instantiated or obtained by the user (as opposed to generating a graph using a mesh and BVH).