Cleaning up compiler warnings as much as possible.
Changed enum COST_ALG_KEY to const enum class COST_ALG_KEY. This should be changed in the actual codebase, as it is safer than a C-style enum.

Also, std::pow returns a double, but what if the template parameter for euclidean_distance is float? (It is a non catastrophic warning, but it is preferable to eliminate warnings whenever possible.) I do a static_cast here to ensure that std::pow is cast to whatever the user's template parameter is.

The whole goal here was to clean up all of the compiler warnings.